refactor!: перемещение файлов в правильную дерикторию
