+++ app.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"connect-text-bot/bot"
-	"connect-text-bot/botconfig_parser"
-	"connect-text-bot/config"
-	"connect-text-bot/database"
-	"connect-text-bot/logger"
+	"connect-text-bot/internal/botconfig_parser"
+	"connect-text-bot/internal/config"
+	"connect-text-bot/internal/database"
+	"connect-text-bot/internal/logger"
 	"context"
 	"flag"
 	"log"
