+++ bot/client/client.go
@@ -11,8 +11,8 @@ import (
 	"time"
 
 	"connect-text-bot/bot/requests"
-	"connect-text-bot/config"
-	"connect-text-bot/logger"
+	"connect-text-bot/internal/config"
+	"connect-text-bot/internal/logger"
 
 	"github.com/google/uuid"
 )
