+++ bot/messages/cache.go
@@ -2,9 +2,9 @@ package messages
 
 import (
 	"connect-text-bot/bot/requests"
-	"connect-text-bot/botconfig_parser"
-	"connect-text-bot/database"
-	"connect-text-bot/logger"
+	"connect-text-bot/internal/botconfig_parser"
+	"connect-text-bot/internal/database"
+	"connect-text-bot/internal/logger"
 	"encoding/json"
 	"errors"
 	"fmt"
